requirements.txt
influxdb-client
ortools
paho-mqtt
flask
request
